move polls outside of clubs domain
