+++ app/polls/apis.py
@@ -1,7 +1,7 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
-from clubs.polls.viewsets import PollViewset
+from polls.viewsets import PollViewset
 
 router = DefaultRouter()
 router.register("polls", PollViewset, basename="polls")
