+++ app/polls/templates/clubs/polls/poll_success.html
