+++ app/polls/tests/test_poll_models.py
@@ -1,4 +1,4 @@
-from clubs.polls.models import Poll, PollField, PollInputType, PollQuestion, TextInput
+from polls.models import Poll, PollField, PollInputType, PollQuestion, TextInput
 from core.abstracts.tests import TestsBase
 from lib.faker import fake
