+++ app/polls/viewsets.py
@@ -1,5 +1,5 @@
-from clubs.polls.models import Poll
-from clubs.polls.serializers import PollSerializer
+from polls.models import Poll
+from polls.serializers import PollSerializer
 from core.abstracts.viewsets import ModelViewSetBase
